+++ src/lib.rs
@@ -1,5 +1,4 @@
 pub mod cmd;
 pub mod legacy;
-pub mod utils;
 pub mod ipfs;
 pub mod utils;
